+++ js/packages/openinference-instrumentation-anthropic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @arizeai/openinference-instrumentation-anthropic
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [4d72f42]
+  - @arizeai/openinference-core@2.6.1
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [99f6e71]
+  - @arizeai/openinference-core@2.6.0
+
 ## 0.2.1
 
 ### Patch Changes