+++ js/packages/openinference-instrumentation-bedrock-agent-runtime/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @arizeai/openinference-instrumentation-bedrock-agent-runtime
 
+## 1.1.24
+
+### Patch Changes
+
+- Updated dependencies [4d72f42]
+  - @arizeai/openinference-core@2.6.1
+
+## 1.1.23
+
+### Patch Changes
+
+- Updated dependencies [99f6e71]
+  - @arizeai/openinference-core@2.6.0
+
 ## 1.1.22
 
 ### Patch Changes