+++ js/packages/openinference-instrumentation-openai-agents/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @arizeai/openinference-instrumentation-openai-agents
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies [4d72f42]
+  - @arizeai/openinference-core@2.6.1
+
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [99f6e71]
+  - @arizeai/openinference-core@2.6.0
+
 ## 0.2.9
 
 ### Patch Changes